Move json-schema-test-suite test to own file
And clean it up a bit.
